0.0.16

+++ src/VCTUp.jl
@@ -1,7 +1,7 @@
 function upgradePCVCT(from_version::VersionNumber, to_version::VersionNumber, auto_upgrade::Bool)
     println("Upgrading pcvct from version $(from_version) to $(to_version)...")
     milestone_versions = [v"0.0.1", v"0.0.3", v"0.0.10", v"0.0.11", v"0.0.13", v"0.0.15"]
-    next_milestone_inds = findall(x -> from_version < x, milestone_versions) # this could be simplified to take advantage of this list being sorted, but who cares? It's already so fast
+    next_milestone_inds = findall(x -> from_version < x, milestone_versions) #! this could be simplified to take advantage of this list being sorted, but who cares? It's already so fast
     next_milestones = milestone_versions[next_milestone_inds]
     success = true
     for next_milestone in next_milestones
@@ -26,7 +26,7 @@ end
 
 function populateTableOnFeatureSubset(db::SQLite.DB, source_table::String, target_table::String; column_mapping::Dict{String, String}=Dict{String,String}())
     source_columns = queryToDataFrame("PRAGMA table_info($(source_table));") |> x -> x[!, :name]
-    target_columns = [c in keys(column_mapping) ? column_mapping[c] : c for c in source_columns]
+    target_columns = [haskey(column_mapping, c) ? column_mapping[c] : c for c in source_columns]
     insert_into_cols = "(" * join(target_columns, ",") * ")"
     select_cols = join(source_columns, ",")
     query = "INSERT INTO $(target_table) $(insert_into_cols) SELECT $(select_cols) FROM $(source_table);"
@@ -37,10 +37,10 @@ function upgradeToV0_0_1(::Bool)
     println("\t- Upgrading to version 0.0.1...")
     data_dir_contents = readdir(joinpath(data_dir, "inputs"); sort=false)
     if "rulesets_collections" in data_dir_contents
-        rulesets_collections_folders = readdir(joinpath(data_dir, "inputs", "rulesets_collections"); sort=false) |> filter(x -> isdir(joinpath(data_dir, "inputs", "rulesets_collections", x)))
-        for rulesets_collection_folder in rulesets_collections_folders
-            path_to_rulesets_collections_folder = joinpath(data_dir, "inputs", "rulesets_collections", rulesets_collection_folder)
-            path_to_rulesets_variations_db = joinpath(path_to_rulesets_collections_folder, "rulesets_variations.db")
+        rulesets_collection_folders = readdir(locationPath(:rulesets_collection); sort=false) |> filter(x -> isdir(locationPath(:rulesets_collection, x)))
+        for rulesets_collection_folder in rulesets_collection_folders
+            path_to_rulesets_collection_folder = locationPath(:rulesets_collection, rulesets_collection_folder)
+            path_to_rulesets_variations_db = joinpath(path_to_rulesets_collection_folder, "rulesets_variations.db")
             if !isfile(joinpath(path_to_rulesets_variations_db))
                 continue
             end
@@ -51,14 +51,14 @@ function upgradeToV0_0_1(::Bool)
             end
             column_names = queryToDataFrame("PRAGMA table_info(rulesets_variations);"; db=db_rulesets_variations) |> x -> x[!, :name]
             filter!(x -> x != "rulesets_collection_variation_id", column_names)
-            path_to_xml = joinpath(path_to_rulesets_collections_folder, "base_rulesets.xml")
+            path_to_xml = joinpath(path_to_rulesets_collection_folder, "base_rulesets.xml")
             if !isfile(path_to_xml)
-                writeRules(path_to_xml, joinpath(path_to_rulesets_collections_folder, "base_rulesets.csv"))
+                writeRules(path_to_xml, joinpath(path_to_rulesets_collection_folder, "base_rulesets.csv"))
             end
             xml_doc = openXML(path_to_xml)
             for column_name in column_names
                 xml_path = columnNameToXMLPath(column_name)
-                base_value = getField(xml_doc, xml_path)
+                base_value = getContent(xml_doc, xml_path)
                 query = "UPDATE rulesets_variations SET '$(column_name)'=$(base_value) WHERE rulesets_collection_variation_id=0;"
                 DBInterface.execute(db_rulesets_variations, query)
             end
@@ -85,7 +85,7 @@ function upgradeToV0_0_3(auto_upgrade::Bool)
         return false
     end
     println("\t- Upgrading to version 0.0.3...")
-    # first get vct.db right changing simulations and monads tables
+    #! first get vct.db right changing simulations and monads tables
     if DBInterface.execute(db, "SELECT 1 FROM pragma_table_info('simulations') WHERE name='config_variation_id';") |> DataFrame |> isempty
         DBInterface.execute(db, "ALTER TABLE simulations RENAME COLUMN variation_id TO config_variation_id;")
     end
@@ -102,27 +102,27 @@ function upgradeToV0_0_3(auto_upgrade::Bool)
         DBInterface.execute(db, "UPDATE monads_temp SET ic_cell_variation_id=CASE WHEN ic_cell_id=-1 THEN -1 ELSE 0 END;")
         DBInterface.execute(db, "DROP TABLE monads;")
         createPCVCTTable("monads", monadsSchema())
-        # drop the previous unique constraint on monads
-        # insert from monads_temp all values except ic_cell_variation_id (set that to -1 if ic_cell_id is -1 and to 0 if ic_cell_id is not -1)
+        #! drop the previous unique constraint on monads
+        #! insert from monads_temp all values except ic_cell_variation_id (set that to -1 if ic_cell_id is -1 and to 0 if ic_cell_id is not -1)
         populateTableOnFeatureSubset(db, "monads_temp", "monads")
         DBInterface.execute(db, "DROP TABLE monads_temp;")
     end
 
-    # now get the config_variations.db's right
+    #! now get the config_variations.db's right
     config_folders = queryToDataFrame(constructSelectQuery("configs"; selection="folder_name")) |> x -> x.folder_name
     for config_folder in config_folders
-        path_to_config_folder = joinpath(data_dir, "inputs", "configs", config_folder)
+        path_to_config_folder = locationPath(:config, config_folder)
         if !isfile(joinpath(path_to_config_folder, "variations.db"))
             continue
         end
-        # rename all "variation" to "config_variation" in filenames and in databases
+        #! rename all "variation" to "config_variation" in filenames and in databases
         old_db_file = joinpath(path_to_config_folder, "variations.db")
         db_file = joinpath(path_to_config_folder, "config_variations.db")
         if isfile(old_db_file)
             mv(old_db_file, db_file)
         end
         db_config_variations = db_file |> SQLite.DB
-        # check if variations is a table name in the database
+        #! check if variations is a table name in the database
         if DBInterface.execute(db_config_variations, "SELECT name FROM sqlite_master WHERE type='table' AND name='variations';") |> DataFrame |> x -> (length(x.name)==1)
             DBInterface.execute(db_config_variations, "ALTER TABLE variations RENAME TO config_variations;")
         end
@@ -227,7 +227,7 @@ function upgradeToV0_0_13(::Bool)
     end
     rulesets_collection_folders = queryToDataFrame(constructSelectQuery("rulesets_collections"; selection="folder_name")) |> x -> x.folder_name
     for rulesets_collection_folder in rulesets_collection_folders
-        path_to_rulesets_collection_folder = joinpath(data_dir, "inputs", "rulesets_collections", rulesets_collection_folder)
+        path_to_rulesets_collection_folder = locationPath(:rulesets_collection, rulesets_collection_folder)
         path_to_new_db = joinpath(path_to_rulesets_collection_folder, "rulesets_collection_variations.db")
         if isfile(path_to_new_db)
             continue
@@ -265,7 +265,7 @@ function upgradeToV0_0_15(auto_upgrade::Bool)
     end
     println("\t- Upgrading to version 0.0.15...")
 
-    # first include ic_ecm_variation_id in simulations and monads tables
+    #! first include ic_ecm_variation_id in simulations and monads tables
     if DBInterface.execute(db, "SELECT 1 FROM pragma_table_info('simulations') WHERE name='ic_ecm_variation_id';") |> DataFrame |> isempty
         DBInterface.execute(db, "ALTER TABLE simulations ADD COLUMN ic_ecm_variation_id INTEGER;")
         DBInterface.execute(db, "UPDATE simulations SET ic_ecm_variation_id=CASE WHEN ic_ecm_id=-1 THEN -1 ELSE 0 END;")
@@ -280,7 +280,7 @@ function upgradeToV0_0_15(auto_upgrade::Bool)
         DBInterface.execute(db, "DROP TABLE monads_temp;")
     end
 
-    # now add ic_dc_id to simulations and monads tables
+    #! now add ic_dc_id to simulations and monads tables
     if DBInterface.execute(db, "SELECT 1 FROM pragma_table_info('simulations') WHERE name='ic_dc_id';") |> DataFrame |> isempty
         DBInterface.execute(db, "ALTER TABLE simulations ADD COLUMN ic_dc_id INTEGER;")
         DBInterface.execute(db, "UPDATE simulations SET ic_dc_id=-1;")
@@ -289,6 +289,47 @@ function upgradeToV0_0_15(auto_upgrade::Bool)
         DBInterface.execute(db, "ALTER TABLE monads ADD COLUMN ic_dc_id INTEGER;")
         DBInterface.execute(db, "CREATE TABLE monads_temp AS SELECT * FROM monads;")
         DBInterface.execute(db, "UPDATE monads_temp SET ic_dc_id=-1;")
+        DBInterface.execute(db, "DROP TABLE monads;")
+        createPCVCTTable("monads", monadsSchema())
+        populateTableOnFeatureSubset(db, "monads_temp", "monads")
+        DBInterface.execute(db, "DROP TABLE monads_temp;")
+    end
+    return true
+end
+
+function upgradeToV0_0_16(auto_upgrade::Bool)
+    warning_msg = """
+    \t- Upgrading to version 0.0.16...
+    \nWARNING: Upgrading to version 0.0.16 will change the database schema.
+    See info at https://drbergman.github.io/pcvct/stable/misc/database_upgrades/
+
+    ------IF ANOTHER INSTANCE OF PCVCT IS USING THIS DATABASE, PLEASE CLOSE IT BEFORE PROCEEDING.------
+
+    Continue upgrading to version 0.0.16? (y/n):
+    """
+    println(warning_msg)
+    response = auto_upgrade ? "y" : readline()
+    if response != "y"
+        println("Upgrade to version 0.0.16 aborted.")
+        return false
+    end
+    println("\t- Upgrading to version 0.0.16...")
+
+    #! add intracellular_id and intracellular_variation_id to simulations and monads tables
+    if DBInterface.execute(db, "SELECT 1 FROM pragma_table_info('simulations') WHERE name='intracellular_id';") |> DataFrame |> isempty
+        DBInterface.execute(db, "ALTER TABLE simulations ADD COLUMN intracellular_id INTEGER;")
+        DBInterface.execute(db, "UPDATE simulations SET intracellular_id=-1;")
+        DBInterface.execute(db, "ALTER TABLE simulations ADD COLUMN intracellular_variation_id INTEGER;")
+        DBInterface.execute(db, "UPDATE simulations SET intracellular_variation_id=-1;")
+    end
+    if DBInterface.execute(db, "SELECT 1 FROM pragma_table_info('monads') WHERE name='intracellular_id';") |> DataFrame |> isempty
+        DBInterface.execute(db, "ALTER TABLE monads ADD COLUMN intracellular_id INTEGER;")
+        DBInterface.execute(db, "UPDATE monads SET intracellular_id=-1;")
+        DBInterface.execute(db, "ALTER TABLE monads ADD COLUMN intracellular_variation_id INTEGER;")
+        DBInterface.execute(db, "UPDATE monads SET intracellular_variation_id=-1;")
+        DBInterface.execute(db, "CREATE TABLE monads_temp AS SELECT * FROM monads;")
+        DBInterface.execute(db, "UPDATE monads_temp SET intracellular_id=-1;")
+        DBInterface.execute(db, "UPDATE monads_temp SET intracellular_variation_id=-1;")
         DBInterface.execute(db, "DROP TABLE monads;")
         createPCVCTTable("monads", monadsSchema())
         populateTableOnFeatureSubset(db, "monads_temp", "monads")
